Bug fix regarding the draw mechanic
For some reason, scripts got stuck on some parts where it had to wait for certain condition. Like wait until the card lerp animation is done. Bool got checkmark for yes, but script did not continue.

Solution for the bug for now at least was just to remove Ienumerators and replace them as normal functions without waiting. Not as fancy, but correct things happen this way now.