Path updating
Finished rest of the paths to lead to end boss as well.
Implemented a rest function to be used in rest points (This opens a new canvas with two options, one to heal and one to increase max hp).
Also started with event canvas, got the layout pretty much done, no functionality for that yet