Removed ap recovery compnent
After some playtesting, decided to remove entire component. It was too good if player could place it in any card. It was stupid not to use it, this removed any challenge or thinking from the game.
Instead decided to use suggestion from Joonas, and use old Breather cards that are always on the deck. 5 seemed a very balanced amount and it did not feel like it had an overpowered effect.